extracted equality operations in a separated module
