+++ lib/mongoid.rb
@@ -48,6 +48,7 @@ require "mongoid/callbacks"
 require "mongoid/config"
 require "mongoid/contextual"
 require "mongoid/copyable"
+require "mongoid/equality"
 require "mongoid/criteria"
 require "mongoid/dirty"
 require "mongoid/factory"
